Create cst
Give a continuous function for implementing the contrast stretching transformation shown in Fig. 3.2(a) (3rd edition). In addition to k, your function must include a parameter, E, for controlling the slope of the function as it transitions from low to high gray-level values. Your function should be normalized (between 0 and 1 or 0 and 255)