Create histogram-equalization-mri
Download Fig. 3.8(a) and perform histogram equalization on it (the MRI of a fractured human spine).
As a minimum, your project solution should include the original image, a plot of its histogram, a plot of the histogram-equalization transformation function, the enhanced image, and a plot of its histogram. Use this information to explain why the resulting image was enhanced as it was.