Create 3x3-median-filter
Write a computer program that will denoise an image using the 3x3 median filter. Apply your algorithm to the X-Ray image of a circuit board corrupted by salt-and-pepper noise (Fig3.37(a).jpg). You should turn in the details of the method, your computer program, the input and output images. For simplicity, perform your calculations only for interior pixels, not for boundary pixels. Explain your result and compare it with the output obtained using the linear average filter.