Create commposite-laplacian
(Composite Laplacian Mask). Write a computer program that implements the operation g(x, y) = f(x, y) − ∇2f(x, y) in the form of a spatial linear filter with a 3x3 mask. Give the form of the mask and apply the program to the image of the North Pole of the Moon (Fig3.40(a).jpg). You should turn in the details of the method, the computer program, the input and output images. Perform your calculations only for interior pixels, not for boundary pixels. Explain your result.